Improved docs, removed useless module, made sys modules pub (crate)
