+++ src/utils/functions.rs
@@ -1,13 +1,11 @@
-use super::sys::winapi::ansi::set_virtual_terminal_processing;
+use super::sys::windows::set_virtual_terminal_processing;
 use lazy_static::lazy_static;
 
 lazy_static! {
     static ref SUPPORTS_ANSI_ESCAPE_CODES: bool = {
-        let is_supported;
-
         // Some terminals on windows like GitBash can't use WinaApi calls directly so when we try to enable the ANSI-flag for windows this won't work.
         // Because of that we should check first if the TERM-variable is set and see if the current terminal is a terminal who does support ANSI.
-        is_supported = if is_specific_term() {
+        let is_supported = if is_specific_term() {
             true
         } else {
             // if it is not listed we should try with WinApi to check if we do support ANSI-codes.
